dfse1: Add nodes automatically when needed
Add a node only if dst or src are not in the range of the graph. Nodes
are numeroted in order 0, 1, ... If we want to use others names like
"node1" we will need to add a hash map that will do the translation
between the node name and its value in the graph. But currently it is ok
to use 0, 1, ...