dfse1: Completed
To complete the task on codingame we just use a list of gateway
and run the BFS for each gateway. We kept track of the closest path
between the viruses and the gateway and cut the link the closest to
the gateway that is on the closest path. It works well. Note that for
distance and probably in some other places we should have use an
Option because we don't check index in our code. But the main purpose
was to experiment the BFS so let's try another challenge now.