calm: Prepare bois1
We add "croissant" and we also modify how state are created. For command
that requires multiple steps we put the plate, do steps and finally take
the plate. By doing like this we can chopped strawberries when ever we
want and not only as the first part of the command.